Fix qtwebengine-everywhere-5.15.5-use-python2.patch.
Fix a missing place where python should be replaced by python2 or
python3. Use python2 for now because everything else does, but this
particular tool actually already supports Python 3.

For some reason, this does not trigger an error in the Fedora buildroot.
(The BuildRequires: /usr/bin/python is not in the Fedora
qt5-qtwebengine.spec, that is why I removed it from here, uncovering
this error.)